+++ src/lib.rs
@@ -1,10 +1,11 @@
 use bevy::prelude::*;
 
+#[cfg(feature = "debug")]
 use crate::debug::DebugPlugin;
 use crate::events::*;
 use crate::resources::*;
 
-pub mod boids;
+pub mod fvo;
 mod cell;
 pub mod components;
 pub mod debug;
@@ -15,7 +16,7 @@ pub mod grid_direction;
 pub mod resources;
 pub mod utils;
 
-use boids::BoidsPlugin;
+use fvo::FvoPlugin;
 use flowfield::FlowfieldPlugin;
 use grid::GridPlugin;
 use resources::ResourcesPlugin;
@@ -25,7 +26,7 @@ pub struct BevyPathfindingPlugin;
 impl Plugin for BevyPathfindingPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins((
-            BoidsPlugin,
+            FvoPlugin,
             FlowfieldPlugin,
             ResourcesPlugin,
             GridPlugin,